Hugely better hash function, overall 2x speedup
Cyclic install computes cyclic hashes much better.  Computes a per-SCC hash, and then a recursive decent DAG hash.  Requires a cycle-leader/color notion (not used by cycle_equals but should be).  Requires a cycle-hash, probably can fold into _hash.
Hash is computed as a long, folded to an int in hashCode.
Type walkers for all types, not just Cyclic.

Rdual just fills in reverse fields, no recursion.

Remove unchecked warnings.
Better Bits hash.

+++ src/main/java/com/cliffc/aa/node/PrimNode.java
@@ -207,7 +207,7 @@ public abstract class PrimNode extends Node {
   // Prims are equal for same-name-same-signature (and same inputs).
   // E.g. float-minus of x and y is NOT the same as int-minus of x and y
   // despite both names being '-'.
-  @Override public int hashCode() { return super.hashCode()+_name.hashCode()+_formals._hash; }
+  @Override public int hashCode() { return super.hashCode()+_name.hashCode()+(int)_formals._hash; }
   @Override public boolean equals(Object o) {
     if( this==o ) return true;
     if( !super.equals(o) ) return false;
